fix: separate code base into modules
